feat: 루트 페이지에 랜딩과 홈 대시보드 병합
