+++ src/pages/home/index.ts
@@ -1,1 +1,2 @@
 export { HomePage } from "./ui/HomePage";
+export { HomeContent } from "./ui/HomeContent";
